+++ src/main/framework/system-commands/reload.ts
@@ -3,7 +3,7 @@ import { ExecuteContext } from '../execute-context'
 
 async function reload(context: ExecuteContext): Promise<void> {
   await context.workspace.load()
-  context.out('- settings reloaded \n')
+  context.out('- settings and themes reloaded \n')
 
   await context.workspace.commands.load(context.workspace.settings)
   context.out('- commands reloaded \n')