+++ src/main/framework/system-commands/restart.ts
@@ -1,0 +1,11 @@
+import { app } from 'electron'
+
+async function restart(): Promise<void> {
+  app.relaunch()
+  app.exit()
+}
+
+export default {
+  command: ':restart',
+  task: restart
+}