+++ src/main/framework/transformers/echo.ts
@@ -1,0 +1,4 @@
+import { transformer } from './index'
+export const echo: transformer = (_, ...args) => {
+  return `echo "${args.join(':')}"`
+}