+++ src/main/framework/transformers/snake.ts
@@ -1,0 +1,4 @@
+import { snakeCase } from 'lodash'
+import { transformer } from './index'
+
+export const snake: transformer = (_, text) => snakeCase(text)