+++ src/main/framework/transformers/upper.ts
@@ -1,0 +1,4 @@
+import { transformer } from './index'
+export const upper: transformer = (_, text) => {
+  return text.toUpperCase()
+}