+++ src/renderer/src/runner/runner.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, ReactElement, useEffect, useRef, useState } from 'react'
+import { ChangeEvent, ReactElement, useEffect, useRef, useState, RefObject, createRef } from 'react'
 import { Command, Runtime } from './runtime'
 import { ContextMenu, ContextMenuItem, ContextMenuTrigger } from 'rctx-contextmenu'
 import CodeMirror from '@uiw/react-codemirror'
@@ -6,13 +6,22 @@ import { vscodeDark } from '@uiw/codemirror-theme-vscode'
 import { color } from '@uiw/codemirror-extensions-color'
 import { hyperLink } from '@uiw/codemirror-extensions-hyper-link'
 import { javascript } from '@codemirror/lang-javascript'
+import RunnerAC from './runner-ac'
+import { Suggestion } from './autocomplete'
 export default function Runner(): ReactElement {
   const [runtimeList, setRuntimes] = useState<Runtime[]>([])
   const [pendingTitles, setPendingTitles] = useState<object>({})
   const [historyIndex, setHistoryIndex] = useState<number>(-1)
   const [commanderMode, setCommanderMode] = useState<boolean>(false)
   const [editMode, setEditMode] = useState<boolean>(false)
+  const [suggestion, setSuggestion] = useState<Suggestion>({
+    list: []
+  })
+  const [suggestionSelection, setSuggestionSelection] = useState<number>(0)
+
   const inputRef = useRef<HTMLInputElement>(null)
+  const historyRefs = useRef<Array<RefObject<HTMLDivElement>>>([])
+
   const textAreaRef = useRef<HTMLTextAreaElement>(null)
 
   const [isMultiLine, setIsMultiLine] = useState<boolean>(false)
@@ -45,6 +54,10 @@ export default function Runner(): ReactElement {
   const runtime = runtimeList.find((runtime) => runtime.target)
   const historicalExecution = historyIndex != -1 ? runtime?.history[historyIndex] : undefined
 
+  historyRefs.current = runtime?.history
+    ? runtime.history.map((_, i) => historyRefs.current[i] ?? createRef())
+    : []
+
   const execute = async (runtime): Promise<void> => {
     const command: Command = await window.electron.ipcRenderer.invoke(
       'runtime.prepareExecute',
@@ -54,6 +67,10 @@ export default function Runner(): ReactElement {
     )
 
     applyHistoryIndex(-1)
+    setSuggestion({
+      list: []
+    })
+
     await reloadRuntimesFromBackend()
 
     // renderer -> "backend"
@@ -73,6 +90,18 @@ export default function Runner(): ReactElement {
 
     await reloadRuntimesFromBackend()
   }
+
+  const changePrompt = (prompt: string): void => {
+    setPrompt(prompt)
+
+    window.electron.ipcRenderer.send('runtime.prompt', prompt)
+
+    const cursor = inputRef?.current?.selectionEnd
+    window.electron.ipcRenderer.invoke('runtime.complete', prompt, cursor ?? -1).then((r) => {
+      setSuggestion(r)
+      setSuggestionSelection(0)
+    })
+  }
   const handlePromptChange = (
     event: ChangeEvent<HTMLInputElement> | ChangeEvent<HTMLTextAreaElement>
   ): void => {
@@ -81,9 +110,7 @@ export default function Runner(): ReactElement {
       applyHistoryIndex(-1)
     }
 
-    setPrompt(value)
-
-    window.electron.ipcRenderer.send('runtime.prompt', value)
+    changePrompt(value)
   }
 
   const handleTitleChange = (id: string, event: ChangeEvent<HTMLInputElement>): void => {
@@ -116,6 +143,15 @@ export default function Runner(): ReactElement {
   const applyHistoryIndex = (index: number): void => {
     window.electron.ipcRenderer.invoke('runtime.reset-focus', runtime?.id).then(() => {
       setHistoryIndex(index)
+      // scroll to history index
+      const element = historyRefs.current[index]?.current
+      if (element) {
+        element.scrollIntoView({
+          behavior: 'smooth',
+          block: 'nearest',
+          inline: 'start'
+        })
+      }
     })
   }
 
@@ -155,12 +191,57 @@ export default function Runner(): ReactElement {
     if (e.code === 'ArrowDown') {
       if (runtime && historyIndex > -1) {
         applyHistoryIndex(historyIndex - 1)
+      } else if (historyIndex === -1) {
+        if (suggestionSelection < suggestion.list.length - 1) {
+          setSuggestionSelection(suggestionSelection + 1)
+        }
+      }
+    }
+
+    if (e.code === 'ArrowLeft' || e.code === 'ArrowRight') {
+      let cursor = inputRef?.current?.selectionEnd ?? -1
+
+      console.log(cursor, runtime?.prompt.length)
+
+      // the selectionEnd hasn't updated yet. add and go
+      if (e.code === 'ArrowLeft') {
+        cursor--
+      } else if (e.code === 'ArrowRight') {
+        cursor++
+      }
+
+      if (cursor <= 0) {
+        cursor = 0
+      } else if (cursor >= (runtime?.prompt?.length ?? 0)) {
+        cursor = runtime?.prompt?.length ?? 0
+      }
+
+      window.electron.ipcRenderer
+        .invoke('runtime.complete', runtime?.prompt, cursor ?? -1)
+        .then((r) => {
+          setSuggestion(r)
+          setSuggestionSelection(0)
+        })
+    }
+
+    if (e.code === 'Tab') {
+      e.preventDefault()
+      e.stopPropagation()
+
+      const suggestionEntry = suggestion.list[suggestionSelection]
+      if (suggestionEntry) {
+        changePrompt(suggestionEntry.prompt)
       }
     }
 
     if (e.code === 'ArrowUp') {
-      if (runtime && historyIndex < runtime.history.length - 1) {
+      if (suggestionSelection > 0) {
+        setSuggestionSelection(suggestionSelection - 1)
+      } else if (runtime && historyIndex < runtime.history.length - 1) {
         applyHistoryIndex(historyIndex + 1)
+        setSuggestion({
+          list: []
+        })
       } else {
         window.electron.ipcRenderer
           .invoke('history.try-scroll-next', runtime?.id)
@@ -296,6 +377,23 @@ export default function Runner(): ReactElement {
   useEffect(() => {
     reloadRuntimesFromBackend().catch((error) => console.error(error))
   }, [])
+
+  useEffect(() => {
+    let styleSheet = document.getElementById('theme')
+    if (!styleSheet) {
+      styleSheet = document.createElement('style')
+      styleSheet.setAttribute('id', 'theme')
+      document.head.appendChild(styleSheet)
+    }
+
+    window.electron.ipcRenderer
+      .invoke('runner.theme', runtimeList.find((o) => o.target)?.profile)
+      .then((theme) => {
+        styleSheet.innerText = theme
+      })
+
+    return () => {}
+  }, [runtimeList])
 
   if (!runtime) {
     return <p>Loading</p>
@@ -435,6 +533,7 @@ export default function Runner(): ReactElement {
               ) : (
                 ''
               )}
+              <RunnerAC suggestion={suggestion} selection={suggestionSelection} />
             </div>
           </div>
           <div className={`runner-result ${result.code !== 0 ? '' : ''}`}>{output}</div>
@@ -444,6 +543,7 @@ export default function Runner(): ReactElement {
             {runtime.history.map((command, index) => (
               <div
                 key={index}
+                ref={historyRefs.current[index]}
                 className={`runner-history-item ${historyIndex === index ? 'runner-history-selected' : ''} ${
                   command.complete
                     ? command.aborted